refactor: QueryModifier 반환타입 추상화
