--- backend/src/main/java/moaon/backend/project/repository/QueryModifier.java
@@ -1,8 +1,0 @@
-package moaon.backend.project.repository;
-
-import moaon.backend.project.dto.ProjectQueryCondition;
-
-public interface QueryModifier<T> {
-
-    T modify(ProjectQueryCondition condition);
-}
