+++ backend/src/main/java/moaon/backend/project/repository/querymodifier/LovesCursorModifier.java
@@ -4,12 +4,12 @@ import static moaon.backend.project.domain.QProject.project;
 
 import com.querydsl.core.BooleanBuilder;
 import lombok.RequiredArgsConstructor;
+import moaon.backend.global.domain.QueryModifier;
 import moaon.backend.global.domain.cursor.Cursor;
 import moaon.backend.project.dto.ProjectQueryCondition;
-import moaon.backend.project.repository.QueryModifier;
 
 @RequiredArgsConstructor
-public class LovesCursorModifier implements QueryModifier<Void> {
+public class LovesCursorModifier implements QueryModifier<Void, ProjectQueryCondition> {
 
     private final BooleanBuilder whereBuilder;
