+++ backend/src/main/java/moaon/backend/project/repository/querymodifier/TechStacksModifier.java
@@ -6,12 +6,12 @@ import com.querydsl.core.BooleanBuilder;
 import com.querydsl.jpa.impl.JPAQuery;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
+import moaon.backend.global.domain.QueryModifier;
 import moaon.backend.project.dto.ProjectQueryCondition;
-import moaon.backend.project.repository.QueryModifier;
 import org.springframework.util.CollectionUtils;
 
 @RequiredArgsConstructor
-public class TechStacksModifier implements QueryModifier<Void> {
+public class TechStacksModifier implements QueryModifier<Void, ProjectQueryCondition> {
 
     private final JPAQuery<?> query;
     private final BooleanBuilder whereBuilder;
